restructure data package
