+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/domain/Result.kt
@@ -1,13 +1,16 @@
-package pw.checkers.data
+package pw.checkers.data.domain
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-enum class PlayerColor() {
+enum class Result {
+    @SerialName("draw")
+    DRAW,
+
     @SerialName("white")
     WHITE,
 
     @SerialName("black")
-    BLACK
+    BLACK,
 }
