+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/request/JoinQueue.kt
@@ -1,7 +1,7 @@
 package pw.checkers.data.request
 
 import kotlinx.serialization.Serializable
-import pw.checkers.data.User
+import pw.checkers.data.domain.User
 
 @Serializable
 data class JoinQueue(
